+++ scrapyd_client/__init__.py
@@ -1,0 +1,3 @@
+from scrapyd_client.pyclient import ScrapydClient
+
+__all__ = [ScrapydClient.__name__]